+++ internal/e2eselfhost/self_host_wasm_component_test.go
@@ -267,10 +267,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -541,10 +541,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -797,10 +797,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_fs(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -1065,10 +1065,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_fs_write(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -1342,10 +1342,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_fs_rw(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -1593,10 +1593,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_random(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -1837,10 +1837,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_env(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -2081,10 +2081,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_args(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -2315,10 +2315,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_clock(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -2551,10 +2551,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_clock_mono(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -2782,10 +2782,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_fs_read_env(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -3029,10 +3029,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_fs_rw_env(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -3270,10 +3270,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_random_write(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -3506,10 +3506,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_eprint(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -3755,10 +3755,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_exit(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -4000,10 +4000,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_fs_args_read(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
@@ -4249,10 +4249,10 @@ function main(): i32 {
         Ok(s) => {
             var core: i32[] = [];
             var i: i32 = 0;
-            while (i < s.len()) { core = core.append(s[i]); i = i + 1; }
+            while (i < s.len()) { core = core.append((s[i] as i32)); i = i + 1; }
             var comp: i32[] = component_full_io_fs_rw_args(core);
             var j: i32 = 0;
-            while (j < comp.len()) { print_int(comp[j]); write("\n"); j = j + 1; }
+            while (j < comp.len()) { print_int((comp[j] as i32)); write("\n"); j = j + 1; }
             return 0;
         },
         Err(e) => { return 1; }
